+++ Afferll/src/Afferll/ToString.h
@@ -5,8 +5,8 @@
 
 namespace Afferll
 {
-	std::string AFRL_API ToString(uint64_t _Val);
-	std::string AFRL_API ToString(int64_t _Val);
+	std::string AFRL_API ToString(unsigned long long _Val);
+	std::string AFRL_API ToString(long long _Val);
 	std::string AFRL_API ToString(float _Val);
 	std::string AFRL_API ToString(double _Val);
 	std::string AFRL_API ToString(bool _Val);
